requirements.txt
langchain_core==0.3.12
langgraph==0.2.39
langchain
langsmith
python-dotenv==1.0.1
SPARQLWrapper==2.0.0
openai
fuzzywuzzy
langdetect
python-Levenshtein
streamlit
langchain-openai